backend/requirements.txt
fastapi
uvicorn
pydantic
dashscope
requests
python-dotenv
openpyxl
google-generativeai
python-jose[cryptography]
passlib[bcrypt]
python-multipart 